issue 5770:  Add ability to assign lot/serial numbers to return authorizations.
git-svn-id: http://svn.code.sf.net/p/postbooks/code/xtuple/trunk@1983 ea1a15ef-c238-0410-93cb-fa28e9f07733

+++ guiclient/returnAuthorization.cpp
@@ -1252,7 +1252,7 @@ void returnAuthorization::closeEvent(QCloseEvent *pEvent)
          (_metrics->value("RANumberGeneration") == "O")   )
       q.prepare("SELECT releaseRaNumber(:number) AS result;");
 
-    q.bindValue(":number", _authNumber->text().toInt());
+    q.bindValue(":number", _authNumber->text());
     q.exec();
     if (q.lastError().type() != QSqlError::None)
     {
